FIX: duplicate in type mapping

+++ startup/20-area-detectors.py
@@ -44,7 +44,7 @@ class TIFFPluginWithFileStore(TIFFPlugin, FileStoreTIFFIterativeWrite):
             raise RuntimeError("SHould never be here")
         
         cam_dtype = self.parent.cam.data_type.get(as_string=True)
-        type_map = {'UInt8': '|u1', 'UInt16': '<u2', 'Float32':'<f4', "Float32":'<f8'}
+        type_map = {'UInt8': '|u1', 'UInt16': '<u2', 'Float32':'<f4', "Float64":'<f8'}
         if cam_dtype in type_map:
             ret[key].setdefault('dtype_str', type_map[cam_dtype])
